+++ Assets/Scripts/ConversationTextManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f1974713e43a4d4bbf0610d02540fa0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
