--- Assets/Scripts/MainTextController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 084ff864ec9e00d41b2cc91484c037e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
