+++ Assets/Scripts/NameTextDrawer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a311cb6d33bfa0b4ab9c95f366b63b55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
